requirement.txt
streamlit
PyPDF2
python-docx
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
